+++ src/components/UserAuthenticator.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
 import logo from '../assets/vectors/logo.svg'
-import Button from './ui/Button'
+import Button from './Button'
 
 enum AuthState {
   loading,
