+++ src/components/__tests__/Icon.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Icon from '../ui/Icon'
+import Icon from '../Icon'
 import { render, cleanup } from 'react-testing-library'
 
 afterEach(() => cleanup())
